Create 2. modular-multiplicative-inverse.rs
Modular multiplicative inverse is an important part of any cryptographic implementation. 

This is my first attempt at writing code in Rust. This program calculates the modular multiplicative inverse of a given number and modulus pair using the Euclidean Algorithm. 

I tried to comment on my code as much as possible so it would provide a lot of explainers to someone who someone experienced in programming but unfamiliar with Rust and the Euclidean Algorithm.